requirement.txt
accelerate==1.7.0
aerich==0.9.0
aiocache==0.12.3
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.17.0
alembic==1.14.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.52.1
anyio==4.9.0
appdirs==1.4.4
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
asyncclick==8.3.0.7
asyncpg==0.31.0
attrs==25.3.0
av==14.4.0
azure-core==1.34.0
azure-storage-blob==12.24.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.14.3
bidict==0.23.1
bitarray==3.4.2
black==25.1.0
blinker==1.9.0
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
click==8.3.1
colbert-ai==0.2.21
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
compressed-rtf==1.0.7
cryptography==46.0.3
ctranslate2==4.6.0
dataclasses-json==0.6.7
datasets==3.6.0
defusedxml==0.7.1
Deprecated==1.2.18
dictdiffer==0.9.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docx2txt==0.8
duckduckgo_search==8.0.2
durationpy==0.10
easygui==0.98.3
ebcdic==1.1.1
ecdsa==0.19.1
einops==0.8.1
elastic-transport==8.17.1
elasticsearch==9.0.1
email-validator==2.3.0
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
Events==0.5
# Editable install with no version control (evoai-backend-template==1.0.0)
-e c:\users\00128738\desktop\ai_code\code_learn_compare_python\fastapi-template-main
extract-msg==0.54.1
fake-useragent==2.1.0
fastapi==0.125.0
faster-whisper==1.1.1
filelock==3.18.0
filetype==1.2.0
firecrawl-py==1.12.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
fpdf2==2.8.2
frozenlist==1.6.0
fs==2.4.16
fsspec==2025.3.0
ftfy==6.2.3
gcp-storage-emulator==2024.8.3
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.15.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
greenlet==3.1.1
grpcio-status==1.67.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
iso8601==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==32.0.1
langdetect==1.0.9
langfuse==2.44.0
lark==1.1.9
ldap3==2.9.1
limits==5.6.0
loguru==0.7.3
lxml==5.4.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
moto==5.1.5
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msoffcrypto-tool==5.4.2
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
oletools==0.60.2
onnxruntime==1.20.1
openai==1.82.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opensearch-py==2.8.0
orjson==3.10.18
overrides==7.7.0
packaging==23.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pcodedmp==1.2.6
peewee==3.18.1
peewee-migrate==1.12.2
pgvector==0.4.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
playwright==1.49.1
pluggy==1.6.0
portalocker==2.10.1
primp==0.15.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.33.2
psutil==7.0.0
psycopg2-binary==2.9.9
py-partiql-parser==0.6.1
pyasn1==0.4.8
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
pyee==12.0.0
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.14.2
pymilvus==2.5.0
pymongo==4.13.0
pypandoc==1.15
pyparsing==3.2.3
pypdf==4.3.1
PyPika==0.48.9
pypika-tortoise==0.5.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.3.5
pytest-docker==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.1
python-iso639==2025.2.18
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-socketio==5.13.0
pytube==15.0.0
pytz==2025.2
pywin32==310
pyxlsb==1.0.10
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
rapidocr-onnxruntime==1.4.4
red-black-tree-mod==1.22
redis==6.2.0
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.7
RestrictedPython==8.0
rsa==4.9.1
RTFDE==0.1.2.1
safetensors==0.5.3
scikit-learn==1.6.1
sentence-transformers==4.1.0
sentencepiece==0.2.0
shapely==2.1.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
starlette==0.48.0
starlette-compress==1.6.0
sympy==1.14.0
tenacity==9.1.2
tencentcloud-sdk-python==3.0.1336
threadpoolctl==3.6.0
tomlkit==0.13.3
tortoise-orm==0.24.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
unstructured==0.16.17
unstructured-client==0.32.3
uritemplate==4.1.1
urllib3==2.6.2
uvicorn==0.38.0
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.4
win32_setctime==1.2.0
win_unicode_console==0.5
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.3
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
